+++ core/common/src/commonMain/kotlin/team/aliens/dms/kmp/core/common/exception/network/ConflictException.kt
@@ -1,4 +1,4 @@
-package team.aliens.dms.kmp.core.network.exception
+package team.aliens.dms.kmp.core.common.exception.network
 
 class ConflictException : NetworkException(
     code = 409,
